requirements.txt
django==4.2.4
gunicorn==21.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2